// frontend/src/lib/openapi/model/aPIKeyScopeType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Teraslack Messaging API
 * Canonical HTTP API contract for the Teraslack server.
 * OpenAPI spec version: 0.1.0
 */

/**
 * Whether the key is user-scoped or workspace-scoped.
 */
export type APIKeyScopeType = typeof APIKeyScopeType[keyof typeof APIKeyScopeType];


export const APIKeyScopeType = {
  user: 'user',
  workspace: 'workspace',
} as const;