// frontend/src/lib/openapi/model/createAPIKeyRequestScopeType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Teraslack Messaging API
 * Canonical HTTP API contract for the Teraslack server.
 * OpenAPI spec version: 0.1.0
 */

/**
 * Whether to create a personal key or a workspace-scoped key.
 */
export type CreateAPIKeyRequestScopeType = typeof CreateAPIKeyRequestScopeType[keyof typeof CreateAPIKeyRequestScopeType];


export const CreateAPIKeyRequestScopeType = {
  user: 'user',
  workspace: 'workspace',
} as const;